Fixed timestamp for archiver in DoocsProcessScalar.
By default, the timestamp that is attached to a variable is the global time stamp, which is only set updated by update() in the adapter, which is not used otherwise.
The adapter already was setting it for the time stamp that is attached to the variable itself, but not for the time stamp that used for archiving. This has been fixed. Now the time stamps in the Doocs histories are set correctly.

To test it run example2 of ApplicationCore, togehter with the oven_sim simulation and watch Oven.heatingCurrent.HIST