Buffer overflow workaround: serverTestSpectrumBuffer
It is possible to set values too quickly into the
referenceTestApplication sender process variables leading to buffer
overflows and loss of information in the receive variablesr queue.

The test was failed intermittently for this reason, A sleep has been
used to slow down the producer as a woe	, but a proper fix would involve
the referenceTestApplication doing its own rate limiting on buffer
overflow situations.