do not send out identical time stamps on the same property (#67)
Identical time stamps are considered to have identical data when DOOCS checks for inconsistencies on silent ZeroMQ connections. If on the same property new data is set without changing the VersionNumber, the timestamp will now be altered by 1 microsecond.

This is more like a temporary work around. A feature request to DOOCS should be made to introduce a distinction between a "server timestamp" and a "source timestamp" like it is realised in OPC UA.

Also a potential better solution to the other part of the problem inside ApplicationCore is described here: https://redmine.msktools.desy.de/issues/9723