Feature: synchronise multi-mapped writeable variables
If a writeable process variable is mapped to multiple properties, so far
the values in the properties were not synchronised, i.e. writing to one
property did not propagate the value to the other connected properties.
This synchronisation is now implemented.

Note that inconsistencies might occur when more than one of the
properties is writeable, hence a warning will be displayed in this case.