fix multi-mapped variable with macro pulse number attached (cf. #9960)
If a macro pulse number is attached to a property which is mapped to a
variable mapped to also other properties, the update from the other,
writeable property was not getting through due to the consistency check.
This check is now effectively disabled in that case, which is ok, since
the update is anyway uncorrelated to the macro pulse number (since
coming from e.g. the operator).