fix crash if property array length mismatch
This can happen e.g. if the ProcessArray length in the application has
changed but the old value is still in the DOOCS config file. If the new
length is bigger than the old one, the server crashed because it
attempted to read too many value from the property in auto_init.

This fix should also cover the case when a wrong length is set trough an
RPC call.