fix: enforce unique timestamp (per property)
The timestamp is used by DOOCS to check whether the last update received
via ZMQ is consistent with the current value (for slowly changing PVs).
Previously it was possible that the timestamp was identical desipte the
value has changed, if the PV changed very fast an odd number of times
with the same timestamp:

1. Very fast changes seem to get lost on ZMQ, only the first update is
transported (despite the latest value ends up in the property and hence
is visible via RPC). This happens also with multiple consecutive
updates.

2. The old mechanism added 1 microsecond, if the timestamp of an update
was identical to the previously sent one, which means the 3rd update
will again get the original timestamp. This renders the DOOCS
consistency check useless if two changes with the same source time stamp
are lost.