fix: increase ZMQ HWM length from 1 to 10
With a value of 1, sending a second value on the same property before
the previous value has been processed (by ZeroMQ) will cause the second
value to be lost. This was observed to happen spuriously in some
cases. The new value should reduce the probability massively, the
remaining cases will be corrected by the DOOCS watchdog generating an
error due to mismatching data.