fix: Propagate initial value for r/o WB variables
If you had a PushInputWB mapped to DOOCS but changed it not to be
writeable, you would not receive a proper initial value but 0.

Now, if there is only one variable connected to this, write down the
initial value nevertheless.

Issue was introduced with 1992ded514439a5e6fb2ab933ec33cb8e8eaf7ee